requirements.txt
# Deep Learning & NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.5.0
keybert>=0.8.0
nltk>=3.8.0
scikit-learn>=1.3.0

# PDF Processing
pymupdf>=1.22.0

# Orchestration
langgraph>=0.1.0
langchain>=0.1.0

# Graph Database
neo4j>=5.14.0

# Vector Database
chromadb>=0.4.0

# LLM Integration
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0

# Data & Utilities
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Retry logic
tenacity>=8.2.0

# Web Interface (Phase 6)
streamlit>=1.28.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
